python-api/requirements.txt
Flask==3.0.3
sentence-transformers==2.6.1
torch>=1.13.1
transformers>=4.25.1
scikit-learn
numpy
python-dotenv

